// src/types/ImageParams.ts
interface ImageParams {
  title?: string;
  technologies?: string;
  subtitleLine1?: string;
  subtitleLine2?: string;
  iconName?: string;
  iconColor?: string;
  titleColor?: string;
  iconUrl?: string;
  iconWidth?: number;
  iconOffsetTop?: number;
  iconOffsetBottom?: number;
  iconOffsetLeft?: number;
  iconOffsetRight?: number;
  fontName?: string;
}

export default ImageParams;